fatboot: feat [XJ3-3974] Improve fastboot to support 64bit app, for big image(< 4G).
Summary:
Improve fastboot to support 64bit app, for big image(< 4G).

It is for fastboot 64bit support.

eg.
fastboot fetch 0x0@userdata disk.img    # which will fetch image bigger than 4G


Test Plan: ok

Reviewers: yadong.li, ming.yu, yan04.zhang, pengfei.xu, yunqian.wang

Reviewed By: yadong.li, yunqian.wang

Subscribers: jianghe.xu

Differential Revision: https://cr.hobot.cc/D152829

+++ include/fb_mmc.h
@@ -44,6 +44,6 @@ void fastboot_mmc_erase(const char *cmd, char *response);
  * @offset: offset that bytes already loaded
  * @response: Pointer to fastboot response buffer
  */
-int fastboot_mmc_flash_read(char *cmd, void *upload_buffer,
-			u32 buffer_size, u32 offset, char *response);
+int64_t fastboot_mmc_flash_read(char *cmd, void *upload_buffer,
+			u64 buffer_size, s64 offset, char *response);
 #endif
